separate movies and shows + rework raiting
