+++ data/fixtures/movie.json
@@ -10,7 +10,7 @@
             "duration": 121,
             "budget": 11000000,
             "revenue": 775340000,
-            "rating": 4.5,
+            "rating": 90,
             "description": 
             "At the height of the Galactic Civil War, Princess Leia have managed to steal the plans to the Death Star, a never-before-seen weapon able to destroy entire planets. In her escape she is taken prisoner by Darth Vader, the infamous leader of the armies of the Galactic Empire, eventually crossing paths with a farmboy called Luke Skywalker, his elderly and mysterious friend Ben Kenobi, and their pilot, the charming but self-centred Han Solo. With a variety of motives, they all agree to help the Princess deliver the Death Star plans to the Resistance, in the hope that any weakness might allow a desperate fight back against the evil of the Empire."
         }
