+++ public/css/stars.css
@@ -1,36 +1,112 @@
-@charset "UTF-8";
-:root {
-  --star-size: 60px;
-  --star-color: #fff;
-  --star-background: #fc0;
-}
+/*
+ * Copyright (c) 2012-2013 Thibaut Courouble
+ * http://www.cssflow.com
+ * Licensed under the MIT License
+ *
+ * Sass/SCSS source: https://goo.gl/UhXKg
+ * PSD by Vin Thomas: https://goo.gl/n1M2e
+ */
+  
+  .container {
+    margin: 0px;
+    width: 250px;
+    text-align: left;
+  }
+  
+  .container .progress {
+    margin: 0 auto;
+    width: 250px;
+  }
+  
+  .progress {
+    padding: 4px;
+    background: rgba(0, 0, 0, 0.25);
+    border-radius: 6px;
+    -webkit-box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.25), 0 1px rgba(255, 255, 255, 0.08);
+    box-shadow: inset 0 1px 2px rgba(0, 0, 0, 0.25), 0 1px rgba(255, 255, 255, 0.08);
+  }
+  
+  .progress-bar {
+    height: 16px;
+    border-radius: 4px;
+      background-image: -webkit-linear-gradient(top, rgba(255, 255, 255, 0.3), rgba(255, 255, 255, 0.05));
+    background-image: -moz-linear-gradient(top, rgba(255, 255, 255, 0.3), rgba(255, 255, 255, 0.05));
+    background-image: -o-linear-gradient(top, rgba(255, 255, 255, 0.3), rgba(255, 255, 255, 0.05));
+    background-image: linear-gradient(to bottom, rgba(255, 255, 255, 0.3), rgba(255, 255, 255, 0.05));
+    -webkit-transition: 0.4s linear;
+    -moz-transition: 0.4s linear;
+    -o-transition: 0.4s linear;
+    transition: 0.4s linear;
+    -webkit-transition-property: width, background-color;
+    -moz-transition-property: width, background-color;
+    -o-transition-property: width, background-color;
+    transition-property: width, background-color;
+    -webkit-box-shadow: 0 0 1px 1px rgba(0, 0, 0, 0.25), inset 0 1px rgba(255, 255, 255, 0.1);
+    box-shadow: 0 0 1px 1px rgba(0, 0, 0, 0.25), inset 0 1px rgba(255, 255, 255, 0.1);
+    text-align: center;
+    font-size: medium;
+    font-weight: bolder;
+    color: black;
+  }
 
-.stars {
-  --percent: calc(var(--rating) / 5 * 100%);
-  display: inline-block;
-  font-size: var(--star-size);
-  font-family: Times;
-  line-height: 1;
-  float: left;
-}
-.stars::before {
-  content: '★★★★★';
-  letter-spacing: 3px;
-  background: linear-gradient(90deg, var(--star-background) var(--percent), var(--star-color) var(--percent));
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  float: left;
-}
+  .progress-bar p {
+      margin-bottom: 0;
+  }
 
-body {
-  background: #eee;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
+  .progress > .progress-bar {
+      /* --wdt:  */
+      --mg: 2
+      
+  }
 
-* {
-  position: relative;
-  box-sizing: border-box;
-}
+  .progress > .progress-bar {
+      width: calc(var(--rating) * 1%);
+      /* background    : linear-gradient(to right, rgba(250, 0, 0) , rgba(250, 0,   0),    rgba(250, 0  , 0));     
+      background    : linear-gradient(to right, rgba(250, 0, 0) , rgba(250, 125, 0),    rgba(250, 125, 0));     
+      background    : linear-gradient(to right, rgba(250, 0, 0) , rgba(250, 125, 0),    rgba(125, 250, 0));     
+      background    : linear-gradient(to right, rgba(250, 0, 0) , rgba(250, 250, 0),    rgba(125, 250, 0));     */
+      background    : linear-gradient(to right, rgba(250, 0, 0) , rgba(250, var(--fm), 0), rgba(var(--sl),   var(--sm), 0));
+
+      /* background    : linear-gradient(to right, rgba(250, 0, 0) , rgba(250, (calc(var(--mg))*125), 0),    rgba(0,   250, 0));        */
+      
+      /*
+            20 -> 0, 40-60 -> 1, 60-100 -> 2
+            1 -> 0, 2-3 -> 1, 4-5 -> 3   100/20 + 1
+
+
+
+      */
+  }
+  
+  /*
+   * Note: using adjacent or general sibling selectors combined with
+   *       pseudo classes doesn't work in Safari 5.0 and Chrome 12.
+   *       See this article for more info and a potential fix:
+   *       https://css-tricks.com/webkit-sibling-bug/
+   */
+  
+  /* #five:checked ~ .progress > .progress-bar {
+    width: 5%;
+    background-color: #f63a0f;
+  }
+  
+  #twentyfive:checked ~ .progress > .progress-bar {
+    width: 25%;
+    background-color: #f27011;
+  }
+  
+  #fifty:checked ~ .progress > .progress-bar {
+    width: 50%;
+    background-color: #f2b01e;
+  }
+  
+  #seventyfive:checked ~ .progress > .progress-bar {
+    width: 75%;
+    background-color: #f2d31b;
+  }
+  
+  #onehundred:checked ~ .progress > .progress-bar {
+    width: 100%;
+    background-color: #86e01e;
+  }
+   */
